Phase 13.25.DF FIX1: AD-52 sentinel + bool fix + test quality
PHASE_13_25_DF_FIX1_END

Implementation fixes:
  I-1: AD-52 error=None sentinel — distinguishes user-explicit error='sem'
       from signature default. Mirrors R2 central=None pattern.
       (Claude49 P1-1 + Claude37 P1-2, convergent)
  I-2: error_bars + error='none' dispatch — was rendering nothing.
       Now renders quantile bars only. (Claude37 P1-1, unique catch)
  I-3: BUG_dfdraw_20260505 — boolean expressions (==,!=,>,<,&,|,~)
       crash np.histogram. histogram.py was the only plot function
       missing .astype(float) at 3 data-entry locations.
  I-4: yerr clamp np.maximum(0,...) — boolean data edge case where
       quantile exceeds central value produces negative delta.

Test quality fixes (Claude48/Claude49/Claude37 convergent):
  T-1: Class 4 — real ErrorbarContainer assertions, capsize independence
  T-2: Class   T-2: Class   T-2: Class   T-2: Class   Tf polys:' guar  T-2: Class   T-2: Class   Tssertions on error kwarg interaction
  T-4: Class 7 — renamed to match actual behavior (_last_ax, not color cycle)
  T-5: Class 10 â  T-5: Class 10 â  T-5: Class 10 â  T-5: Class 10 â  T-5: ClClass 11 — replaced byte_identical with structural equality
  T-7: Classes 14+15 — 11 bool tests through DFDraw (not raw matplotlib)

Process lesson (Reviewer Card v1.28 addition):
  For phases adding >20 new tests, reviewer must sample >=5 test bodies
  and verify assertions match named identities. Capability matrix
  showing 0 invariance is a first-pass quality gate signal.

Test results: 577 passed, 0 failed (485 existing + 92 new)
Capability matrix: 52 features, 0 Broken

+++ UTILS/dfextensions/dfdraw/plots/histogram.py
@@ -238,15 +238,18 @@ def draw_hist(
         fig = ax.get_figure()
     
     # Get data
+    # BUG_dfdraw_20260505: cast to float — boolean expressions (==, !=, >, <, &, |, ~)
+    # produce np.bool_ columns; np.histogram cannot subtract boolean edges.
+    # Profile, hist2d, hexbin all already cast to float here; hist was the outlier.
     if isinstance(x, str):
         x_name = x
-        x_data = df[x].values
+        x_data = df[x].values.astype(float)
     else:
         x_name = "x"
-        x_data = np.asarray(x)
+        x_data = np.asarray(x, dtype=float)
     
     # Remove NaN
-    mask = ~np.isnan(x_data.astype(float))
+    mask = ~np.isnan(x_data)
     x_data = x_data[mask]
     
     # Statistics dict
@@ -357,13 +360,15 @@ def _draw_hist_grouped(
     
     if stacked:
         # Stacked histogram
-        data_list = [df[df[group_by] == g][x].dropna().values for g in groups]
+        # BUG_dfdraw_20260505: cast to float for boolean expressions
+        data_list = [df[df[group_by] == g][x].dropna().values.astype(float) for g in groups]
         ax.hist(data_list, label=[str(g) for g in groups], color=colors,
                 stacked=True, **hist_kwargs)
     else:
         # Overlaid histograms
         for i, group in enumerate(groups):
-            group_data = df[df[group_by] == group][x].dropna().values
+            # BUG_dfdraw_20260505: cast to float for boolean expressions
+            group_data = df[df[group_by] == group][x].dropna().values.astype(float)
             ax.hist(group_data, label=str(group), color=colors[i], **hist_kwargs)
